Saga in subscriber-after uptade BMI and tracking publish sub to measure

+++ Measure.Services/Models/MeasureModel.cs
@@ -8,7 +8,7 @@ namespace Measure.Services.Models
     {
         public int CardId { get; set; }
 
-        public decimal Weight { get; set; }
+        public float Weight { get; set; }
 
         public DateTime Date { get; set; }
